[PORT FROM MAIN] Fix crashes introduced when NUM_BUFFER_SLOTS increased to 64
BZ: 128174

This is to fix the compiling error caused by changing slot from u32
to u64. Details lists below.

Change http://android.intel.com:8080/87673 increased the value of the
constant NUM_BUFFER_SLOTS from 32 to 64. The methods
BufferQueue:getReleasedBuffers and ConsumerBase::onBuffersReleased
were written assuming that NUM_BUFFER_SLOTS is 32 or less. If
NUM_BUFFER_SLOTS is greater than 32, buffers that are still in use are
sometimes freed, resulting in crashes and other catastrophic failures.
I changed those two methods so that they work correctly when
NUM_BUFFER_SLOTS is 64.

Category: aosp improvement
Domain: Graphics-SF
Origin: internal
Upstream-Candidate: no, proprietary

+++ media/libstagefright/omx/GraphicBufferSource.cpp
@@ -722,7 +722,7 @@ void GraphicBufferSource::onFrameAvailable() {
 void GraphicBufferSource::onBuffersReleased() {
     Mutex::Autolock lock(mMutex);
 
-    uint32_t slotMask;
+    uint64_t slotMask;
     if (mBufferQueue->getReleasedBuffers(&slotMask) != NO_ERROR) {
         ALOGW("onBuffersReleased: unable to get released buffer set");
         slotMask = 0xffffffff;
